Split the class course
