--- .setup/bin/sample_courses/models/course/course_create.py
@@ -1,7 +1,0 @@
-"""
-Contains Create_Course class that inheits Course_utils class
-"""
-from .course_utils import Course_utils
-
-class Create_Course(Course_utils):
-    pass
